Completed module

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,28 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, 
+            UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var model = _context.Items.Where(a=> a.User.Id == CurrentUser.Id).ToList();
 
             return View("Index", model);
         }
@@ -29,6 +36,7 @@ namespace WishList.Controllers
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            item.User = CurrentUser;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -37,9 +45,22 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+
+            if(item.User.Id != CurrentUser.Id)
+            {
+                return Unauthorized();
+            }
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
+        }
+
+        private ApplicationUser CurrentUser
+        {
+            get
+            {
+                return _userManager.GetUserAsync(HttpContext.User).Result;
+            }
         }
     }
 }
